+++ fast_rx_persistence/lib/fast_rx_persistence.dart
@@ -2,4 +2,5 @@ library fast_rx_persistence;
 
 export 'src/exceptions.dart';
 export 'src/fast_rx_persistence_base.dart';
+export 'src/fast_rx_persistence_converter.dart';
 export 'src/fast_rx_persistence_interface.dart';