+++ fast_rx_persistence/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+- BREAKING: Replaces `decode` and `encode` parameters with a `converter` paramerter that takes a `FastRxPersistenceConverter`. The simplest migration is to pass in an `InlineConverter` with `fromStore` and `toStore` parameters.
+- Adds `EnumStringConverter` and `EnumIntConverter` for convenience
+
 ## 0.1.5
 - Updated `fast_rx` to 0.4.0